ten_agent/requirements.txt
# TEN Framework
ten_framework>=0.1.0

# AI Models
openai>=1.0.0
aiohttp>=3.9.0

# FunASR/SenseVoice
funasr>=1.0.0
modelscope>=1.10.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
numpy>=1.24.0
soundfile>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
